Move SubsystemDescriptionDump & friends to test code
