+++ host-controller/src/main/java/org/jboss/as/domain/controller/resources/DomainRootDefinition.java
@@ -76,7 +76,6 @@ import org.jboss.as.controller.persistence.ExtensibleConfigurationPersister;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
 import org.jboss.as.controller.services.path.PathManagerService;
 import org.jboss.as.controller.services.path.PathResourceDefinition;
-import org.jboss.as.controller.transform.SubsystemDescriptionDump;
 import org.jboss.as.domain.controller.DomainController;
 import org.jboss.as.domain.controller.HostRegistrations;
 import org.jboss.as.domain.controller.LocalHostControllerInfo;
@@ -285,9 +284,6 @@ public class DomainRootDefinition extends SimpleResourceDefinition {
             resourceRegistration.registerOperationHandler(SnapshotListHandler.DEFINITION, snapshotList);
             SnapshotTakeHandler snapshotTake = new SnapshotTakeHandler(configurationPersister);
             resourceRegistration.registerOperationHandler(SnapshotTakeHandler.DEFINITION, snapshotTake);
-
-            final SubsystemDescriptionDump dumper = new SubsystemDescriptionDump(extensionRegistry);
-            resourceRegistration.registerOperationHandler(SubsystemDescriptionDump.DEFINITION, dumper);
 
         } else {
             DeploymentUploadURLHandler.registerSlave(resourceRegistration);
