+++ subsystem-test/tests/src/test/java/org/jboss/as/subsystem/test/transformers/subsystem/similarity/ModelDescriptionTestCase.java
@@ -1,4 +1,22 @@
-package org.jboss.as.controller.transform;
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2017, Red Hat, Inc., and individual contributors as indicated
+ * by the @authors tag.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.as.subsystem.test.transformers.subsystem.similarity;
 
 import java.util.List;
 
@@ -9,6 +27,7 @@ import org.jboss.as.controller.ProcessType;
 import org.jboss.as.controller.ResourceDefinition;
 import org.jboss.as.controller.registry.LegacyResourceDefinition;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
+import org.jboss.as.subsystem.test.SubsystemDescriptionDump;
 import org.jboss.dmr.ModelNode;
 import org.junit.AfterClass;
 import org.junit.Assert;
@@ -16,6 +35,9 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 /**
+ * This code is a test stub for idea around using {@link SimilarityIndex} to automatically generate transformation rules
+ * Implementation was never completed to the point of using it for subsystem transformation but still exists here if it might be needed in future.
+ *
  * @author <a href="mailto:tomaz.cerar@redhat.com">Tomaz Cerar</a>
  */
 public class ModelDescriptionTestCase {
