+++ subsystem-test/tests/src/test/java/org/jboss/as/subsystem/test/transformers/subsystem/similarity/RootSubsystemResource.java
@@ -1,4 +1,22 @@
-package org.jboss.as.controller.transform;
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2017, Red Hat, Inc., and individual contributors as indicated
+ * by the @authors tag.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.as.subsystem.test.transformers.subsystem.similarity;
 
 import org.jboss.as.controller.NoopOperationStepHandler;
 import org.jboss.as.controller.PathElement;
