+++ subsystem-test/tests/src/test/java/org/jboss/as/subsystem/test/transformers/subsystem/similarity/SessionDefinition.java
@@ -1,15 +1,34 @@
-package org.jboss.as.controller.transform;
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2017, Red Hat, Inc., and individual contributors as indicated
+ * by the @authors tag.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.jboss.as.subsystem.test.transformers.subsystem.similarity;
 
 import org.jboss.as.controller.AttributeDefinition;
 import org.jboss.as.controller.NoopOperationStepHandler;
+import org.jboss.as.controller.OperationDefinition;
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.ReloadRequiredWriteAttributeHandler;
 import org.jboss.as.controller.SimpleAttributeDefinition;
 import org.jboss.as.controller.SimpleAttributeDefinitionBuilder;
+import org.jboss.as.controller.SimpleOperationDefinitionBuilder;
 import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.descriptions.NonResolvingResourceDescriptionResolver;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
-import org.jboss.as.controller.test.TestUtils;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
 
@@ -34,7 +53,7 @@ public class SessionDefinition extends SimpleResourceDefinition {
                     .setAllowExpression(true)
                     .setDefaultValue(null)
                     .setRestartAllServices()
-                    .setAllowNull(true)
+                    .setRequired(false)
                     .build();
     protected static final SimpleAttributeDefinition DEBUG =
             new SimpleAttributeDefinitionBuilder("debug", ModelType.BOOLEAN, true)
@@ -54,6 +73,12 @@ public class SessionDefinition extends SimpleResourceDefinition {
     @Override
     public void registerOperations(ManagementResourceRegistration registry) {
         super.registerOperations(registry);
-        registry.registerOperationHandler(TestUtils.createOperationDefinition("dump-session-info",DEBUG), NoopOperationStepHandler.WITHOUT_RESULT);
+        registry.registerOperationHandler(createOperationDefinition("dump-session-info", DEBUG), NoopOperationStepHandler.WITHOUT_RESULT);
+    }
+
+    private static OperationDefinition createOperationDefinition(String name, AttributeDefinition... parameters) {
+        return new SimpleOperationDefinitionBuilder(name, new NonResolvingResourceDescriptionResolver())
+                .setParameters(parameters)
+                .build();
     }
 }
